backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
torch
scikit-learn
paho-mqtt
pandas
